Add processAliasModes utility function and tests
